Use -Wno-unused-parameter instead of the UNUSED macro.
The UNUSED macro was not used in the code generated by defn2c.pl, so
there were still a lot of warnings during compilation. The only reason
the warnings are there is because the -W option is used, so instead of
hunting down every place where parameters are unused, just disable that
specific warning.