Redo no_act: cleanup in main.c.
The label no_act is misleading; it can be jumped to even if no_act is
false. So rename it to "end", since that is a better description of what
is actually happening.

There is no need to set the FILE * variables to zero at the end of the
function, it is quite clear that they are not going to be used.