Detect recursion.
When ifup/ifdown is running scripts, it could be that those scripts
recursively call ifup/ifdown. If that's for a different interface it's
OK, but if they try to ifup/ifdown an interface that is already in the
process of being brought up/down, then it's an error. To detect this
situation, add environment variables that keep track of which interfaces
are being worked on.

Note that this does not prevent parallel execution of ifup/ifdown of the
same interface.