Per-interface state files with locking.
Instead of having /run/network/ifstate and .ifstate.lock, state is now
maintained in /run/network/ifstate.$iface, and the latter is locked for
the whole time the interface is being acted upon. This prevents two
ifup/ifdown processes from simultaneously acting on the same interface.

If ifup/ifdown tries to lock an already locked statefile, it will print
a warning that it's waiting for the lock, and then wait until it itself
can get a lock. Once it has the lock, it will read the state file to see
if it needs to do anything, if so it will proceed to do so. Only when it
is completely finished with the interface will it release the lock file.

Currently, the /run/network/ifstate file itself will be kept updated for
backwards compatibility with external tools that might use it, but it is
no longer used to determine the interface state.