On ifdown, only remove addresses added by ifup. (Closes: #750740)
For statically configured interfaces, we know exactly which address and
gateway we have given an interface. Instead of running "ip addr flush"
on ifdown, specify the addresses cofigured in /etc/network/interfaces.
This is especially important for IPv6 addresses on alias interfaces,
since the ip command has no way of telling that some addresses only
belong to a specific alias interface.