Make method order deterministic when generating C code. Closes: #762388
As part of the “reproducible builds” effort, we have detected that
`ifupdown` did not build reproducibly.

This is caused by the fact that Perl orders its hashes randomly by
default. This, in turn, results in a random order in which the methods
are written when generating C code. Resulting in different binaries for
each builds.

This commit will sort the methods to produce a stable order.
`ifupdown` can then build reproducibly. :)