Add a systemd service file.
This takes over from the old init file. The only thing it does is call
ifup -a and ifdown -a. It doesn't set up the /run/network directory and
the ifstate file anymore, as ifupdown can now take care of that itself.
It's scheduled to run after systemd-udev-settle, network-pre.target, and
before network-online.target.

The options from /etc/default/networking are no longer honored.