Wait for a link-local address before starting a DHCPv6 client. (Closes: #806673)
On Linux, an interface is only given a link-local IPv6 address when it
a carrier is detected. For a wired Ethernet interface, this may happen
very quickly, but for a wireless interface it may require that the
wireless card associates with a base station first.

DHCPv6 needs a link-local address to function. Unfortunately, the ISC
DHCP client bails out immediately when it doesn't find one, instead of
going to the background and waiting for one to come up. To work around
this problem, call the wait-for-ll6.sh script before starting dhclient
to wait (at most a few seconds) for a link-local address to appear.