Fix FTBFS while running the testsuite. Closes: #807617
For IPv6 stanza's, settle-dad.sh is run, but only if
/lib/ifupdown/settle-dad.sh exists and is executable. The testsuite
expected this to always be the case, but it is not in a completely clean
build environment where ifupdown has not been installed in. So make it
so that settle-dad.sh is never called when running the testsuite, and
update the expected output accordingly.