Fix ifquery crash if interface state file does not exist yet
We do this in do_interface():

	char *current_state;
	lock = lock_interface(iface, &current_state);

current_state is a stack variable and thus could have a random value. For
"ifquery" (or --no-act), lock_interface() does not exit(1) on failure but
returns NULL instead. This happens if if /run/network/ifstate.<iface> does not
exist (yet), then lock == NULL, and current_state does not get written to as
that's an early exit path in lock_interface().

Then do_interface progresses into the "else if (cmds == iface_query)" branch,
current_state != NULL is satisfied, and strncpy() gets called on the
uninitialized value.

Fix this by initializing current_state to NULL.

Closes: #810779
LP: #1532722