ifup@.service: Avoid stopping on shutdown via stopping system-ifup.slice
This is a small behaviour change in systemd 228, as instantiated units now
Require= their parent slice instead of just Wants= it.

We don't need resource control on the group of ifup@ units, so just let them
run in the normal system.slice; this avoids stopping them on shutdown, as this
interferes with NFS root and isn't necessary as networking.service already
takes care of shutting down interfaces.

Closes: #761909
LP: #1492546