Allow ifdown/ifquery when physical interface is not same as logical interface.
Commit 21f2c8863ad7a187877a5be4a4ecaffa12157585 introduced a problem
when doing the following:

ifup foo=bar
ifdown foo

If there is no "iface foo" stanza in /e/n/i, the early check introduced
by that commit will cause ifdown to exit with an error. Fix this by
checking if there is a state file for the physical interface present.